EN6-48 content model list - route structure reverted to old structure but route values remain in route.js
